requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.5
django-appconf==1.0.4
django-celery==3.3.1
django-celery-email==3.0.0
django-celery-results==2.0.1
django-import-export==2.5.0
django-solo==1.1.5
django-tinymce==3.2.0
Django-tinymce-filebrowser==0.2.1
djangorestframework==3.12.2
et-xmlfile==1.1.0
gunicorn==20.0.4
kombu==5.0.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.5
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tablib==3.0.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
